+++ example/src/ColumnProps.tsx
@@ -7,12 +7,12 @@ import { InlineCode } from "./components/library/InlineCode";
 const StyledTable = styled(Table)`
   border: 1px solid #ececec;
 
-  .react-fluid-table-header {
+  .rft-header {
     background-color: #282a36;
   }
 
-  .cell,
-  .header-cell {
+  .rft-cell,
+  .rft-header-cell {
     border-right: 1px solid #ececec;
 
     :last-child {
@@ -20,7 +20,7 @@ const StyledTable = styled(Table)`
     }
   }
 
-  .header-cell-text {
+  .rft-header-cell-text {
     color: #ff79c5;
   }
 `;