+++ CHANGELOG.md
@@ -1,12 +1,21 @@
 # CHANGELOG
 
+## 0.5.3
+
+_2024-06-13_
+
+### Features
+
+- adds `row` and `index` as properties when using `expander`
+- adds `onExpandRow` to allow row expansion to be controlled
+
 ## 0.5.2
 
 _2024-05-30_
 
 ### Features
 
-- adds `rowContainerStyle` vs. `roeContainerClassname`
+- adds `rowContainerStyle` and `rowContainerClassname`
 
 ## 0.5.1